Initial proof of concept.
sysfuck currently only recognises the 'brk' syscall, and its output is
targeted at humans, not Turing tarpits. But it seems to work, at least
on my computer when compiled with gcc. I don't know how many unsafe
assumptions it makes.

bindump.pl is a tool to make it easier to talk to sysfuck. It's like a
reverse hexdump: you give it a sequence of hexadecimal numbers and it
produces binary output. If necessary, a hex number is 0-padded in front
to produce a whole number of bytes; `ab cd` produces the same output as
`abcd`, but `abc def` is `0abc0def` rather than `abcdef`. You can also
output a string of printable characters by enclosing it in single or
double quotes, so `'abc'` is the same as `61 62 63`.

The bindump.pl lexer is a little funky, with the practical consequence
being that output is only sent when a string is read or the input stream
closes. So sending the input
    'foo'
    64
    'foo'
Will result in `foo` being printed after the first line, nothing being
printed after the second, and `dfoo` being printed after the third.