Change semantics: input is passed by default, commands need flagging.
Specifically, most characters sent to sysfuck will by default get
written directly to fd 4 (which sfwrap opens to stdout). This should
make debugging a lot easier, as well as general use.

The exception is a NUL byte, which specifies that a command is to
follow. This works exactly the same as it used to, except that an empty
command (ie. printing NUL NUL) sends a NUL to fd 4.